Refactor kernelspec resource handler to separate URL prefix
